requirements.txt
fastapi~=0.115.12
uvicorn[standard]~=0.34.3
sqlmodel~=0.0.24
python-dotenv
python-chess
celery[redis]~=5.5.1
psycopg[binary]
pydantic~=2.11.3
python-chess
celery
sqlmodel
psycopg[binary]
alembic~=1.16.1
sse-starlette~=2.3.6
sseclient-py
requests~=2.32.3
pandas~=2.3.0
numpy~=2.3.0
scikit-learn
celery_once
redis
stockfish
app~=0.0.1
SQLAlchemy~=2.0.41
colorama~=0.4.6
tabulate~=0.9.0
tqdm~=4.67.1